+++ apps/fabric-website/src/pages/Controls/SeparatorPage/SeparatorPage.tsx
@@ -8,7 +8,9 @@ const baseUrl = 'https://github.com/OfficeDev/office-ui-fabric-react/tree/master
 
 export const SeparatorPage: React.StatelessComponent<IControlsPageProps> = props => {
   const { platform } = props;
-  return <ControlsAreaPage {...props} title="Separator" {...SeparatorPageProps[platform]} otherSections={_otherSections(platform)} />;
+  return (
+    <ControlsAreaPage {...props} title="Separator" {...SeparatorPageProps[platform]} otherSections={_otherSections(platform) as any} />
+  );
 };
 
 function _otherSections(platform: Platforms): IPageSectionProps<Platforms>[] {