+++ apps/fabric-website/src/pages/Styles/Colors/NeutralsPage.tsx
@@ -9,7 +9,9 @@ const neutralColors = require<IColorSwatch[]>('@uifabric/fabric-website/lib/data
 
 export const ColorsNeutralsPage: React.StatelessComponent<IStylesPageProps> = props => {
   const { platform } = props;
-  return <StylesAreaPage {...props} {...ColorsNeutralsPageProps[platform]} otherSections={_otherSections(platform)} />;
+  return (
+    <StylesAreaPage {...props} {...ColorsNeutralsPageProps[platform]} otherSections={_otherSections(platform) as IPageSectionProps[]} />
+  );
 };
 
 function _otherSections(platform: Platforms): IPageSectionProps<Platforms>[] {