+++ packages/charting/src/demo/GettingStartedPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { PageMarkdown } from '@uifabric/example-app-base';
+import { Markdown } from '@uifabric/example-app-base';
 import { mergeStyles } from 'office-ui-fabric-react/lib/Styling';
 
 const rootClass = mergeStyles({
@@ -7,11 +7,11 @@ const rootClass = mergeStyles({
   padding: 40
 });
 
-export const GettingStartedPage: React.StatelessComponent = () => {
+export const GettingStartedPage: React.FunctionComponent = () => {
   return (
     <div className={rootClass}>
       {/* tslint:disable-next-line:deprecation */}
-      <PageMarkdown>{require<string>('!raw-loader!../../README.md')}</PageMarkdown>
+      <Markdown>{require<string>('!raw-loader!../../README.md')}</Markdown>
     </div>
   );
 };