+++ packages/date-time/src/components/Calendar/Calendar.tsx
@@ -1,3 +1,4 @@
+import * as React from 'react';
 import { styled } from '@uifabric/utilities';
 import { ICalendarProps, ICalendarStyleProps, ICalendarStyles } from './Calendar.types';
 import { CalendarBase } from './Calendar.base';
@@ -6,7 +7,7 @@ import { styles } from './Calendar.styles';
 /**
  * Calendar description
  */
-export const Calendar: React.StatelessComponent<ICalendarProps> = styled<ICalendarProps, ICalendarStyleProps, ICalendarStyles>(
+export const Calendar: React.FunctionComponent<ICalendarProps> = styled<ICalendarProps, ICalendarStyleProps, ICalendarStyles>(
   CalendarBase,
   styles,
   undefined,