+++ packages/date-time/src/components/Calendar/CalendarYear/CalendarYear.tsx
@@ -1,3 +1,4 @@
+import * as React from 'react';
 import { ICalendarYearProps, ICalendarYearStyleProps, ICalendarYearStyles } from './CalendarYear.types';
 import { getStyles } from './CalendarYear.styles';
 import { styled } from 'office-ui-fabric-react/lib/Utilities';
@@ -6,7 +7,7 @@ import { CalendarYearBase } from './CalendarYear.base';
 /**
  * CalendarYear description
  */
-export const CalendarYear: React.StatelessComponent<ICalendarYearProps> = styled<
+export const CalendarYear: React.FunctionComponent<ICalendarYearProps> = styled<
   ICalendarYearProps,
   ICalendarYearStyleProps,
   ICalendarYearStyles