+++ packages/experiments/src/components/Button/MenuButton/MenuButton.tsx
@@ -1,16 +1,14 @@
-import * as React from 'react';
 import { createComponent } from '../../../Foundation';
-import { MenuButtonState as state } from './MenuButton.state';
+import { useMenuButtonState as state } from './MenuButton.state';
 import { MenuButtonStyles as styles, MenuButtonTokens as tokens } from './MenuButton.styles';
 import { IMenuButtonProps } from './MenuButton.types';
-import { MenuButtonView as view } from './MenuButton.view';
+import { MenuButtonView } from './MenuButton.view';
 
-export const MenuButton: React.StatelessComponent<IMenuButtonProps> = createComponent({
+export const MenuButton: React.StatelessComponent<IMenuButtonProps> = createComponent(MenuButtonView, {
   displayName: 'MenuButton',
   state,
   styles,
-  tokens,
-  view
+  tokens
 });
 
 export default MenuButton;