+++ packages/experiments/src/utilities/factoryComponents.types.tsx
@@ -14,6 +14,6 @@ import { ISlotProp } from '../Foundation';
 export type ICalloutSlot = ISlotProp<ICalloutProps>;
 export type IContextualMenuSlot = ISlotProp<IContextualMenuProps>;
 export type IIconSlot = ISlotProp<IIconProps, string>;
-export type ILabelSlot = ISlotProp<ILabelProps, React.ReactNode>;
-export type IListSlot = ISlotProp<IListProps, React.ReactNode>;
+export type ILabelSlot = ISlotProp<ILabelProps, string>;
+export type IListSlot = ISlotProp<IListProps>;
 export type IPersonaPresenceSlot = ISlotProp<IPersonaPresenceProps, PersonaPresence>;