+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRowFields.types.ts
@@ -42,19 +42,11 @@ export interface IDetailsRowFieldsProps extends IOverrideColumnRenderProps {
   compact?: boolean;
 
   /**
-   * Whether to show shimmer
-   * @deprecated Use `ShimmeredDetailsList` instead: https://developer.microsoft.com/en-us/fabric#/components/detailslist/shimmer
-   */
-  shimmer?: boolean;
-
-  /**
    * Subset of classnames currently generated in DetailsRow that are used within DetailsRowFields.
    */
   rowClassNames: {
     isMultiline: string;
     isRowHeader: string;
-    shimmerIconPlaceholder: string;
-    shimmer: string;
     cell: string;
     cellPadded: string;
     cellUnpadded: string;