+++ packages/office-ui-fabric-react/src/components/DocumentCard/examples/DocumentCard.Compact.Example.tsx
@@ -9,7 +9,7 @@ import {
   DocumentCardType,
   IDocumentCardActivityPerson
 } from 'office-ui-fabric-react/lib/DocumentCard';
-import { Stack } from 'office-ui-fabric-react/lib/Stack';
+import { Stack, IStackTokens } from 'office-ui-fabric-react/lib/Stack';
 import { getTheme } from 'office-ui-fabric-react/lib/Styling';
 import { TestImages } from '../../../common/TestImages';
 
@@ -28,7 +28,8 @@ export class DocumentCardCompactExample extends React.PureComponent {
         {
           name: 'Revenue stream proposal fiscal year 2016 version02.pptx',
           linkProps: {
-            href: 'http://bing.com'
+            href: 'http://bing.com',
+            target: '_blank'
           },
           previewImageSrc: TestImages.documentPreview,
           iconSrc: TestImages.iconPpt,
@@ -37,7 +38,8 @@ export class DocumentCardCompactExample extends React.PureComponent {
         {
           name: 'New Contoso Collaboration for Conference Presentation Draft',
           linkProps: {
-            href: 'http://bing.com'
+            href: 'http://bing.com',
+            target: '_blank'
           },
           previewImageSrc: TestImages.documentPreviewTwo,
           iconSrc: TestImages.iconPpt,
@@ -46,7 +48,8 @@ export class DocumentCardCompactExample extends React.PureComponent {
         {
           name: 'Spec Sheet for design',
           linkProps: {
-            href: 'http://bing.com'
+            href: 'http://bing.com',
+            target: '_blank'
           },
           previewImageSrc: TestImages.documentPreviewThree,
           iconSrc: TestImages.iconPpt,
@@ -55,7 +58,8 @@ export class DocumentCardCompactExample extends React.PureComponent {
         {
           name: 'Contoso Marketing Presentation',
           linkProps: {
-            href: 'http://bing.com'
+            href: 'http://bing.com',
+            target: '_blank'
           },
           previewImageSrc: TestImages.documentPreview,
           iconSrc: TestImages.iconPpt,
@@ -96,8 +100,10 @@ export class DocumentCardCompactExample extends React.PureComponent {
       }
     };
 
+    const stackTokens: IStackTokens = { childrenGap: 20 };
+
     return (
-      <Stack gap={20}>
+      <Stack tokens={stackTokens}>
         {/* Document preview */}
         <DocumentCard type={DocumentCardType.compact} onClickHref="http://bing.com">
           <DocumentCardPreview previewImages={[previewProps.previewImages[0]]} />