+++ packages/office-ui-fabric-react/src/components/HoverCard/ExpandingCard.ts
@@ -1,3 +1,4 @@
+import * as React from 'react';
 import { styled } from '../../Utilities';
 import { IExpandingCardProps, IExpandingCardStyles, IExpandingCardStyleProps } from './ExpandingCard.types';
 import { getStyles } from './ExpandingCard.styles';