+++ packages/office-ui-fabric-react/src/components/Panel/Panel.styles.ts
@@ -12,6 +12,7 @@ import {
   ScreenWidthMinUhfMobile,
   IStyle
 } from '../../Styling';
+import { FontWeights } from '../../Styling';
 import { getWindow } from '../../Utilities';
 
 // TODO -Issue #5689: Comment in once Button is converted to mergeStyles
@@ -200,7 +201,7 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
     theme,
     type = PanelType.smallFixedFar
   } = props;
-  const { palette } = theme;
+  const { palette, effects } = theme;
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
   const isCustomPanel = type === PanelType.custom || type === PanelType.customNear;
   const win = getWindow();
@@ -243,7 +244,7 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
       classNames.main,
       {
         backgroundColor: palette.white,
-        boxShadow: '0px 0px 30px 0px rgba(0,0,0,0.2)',
+        boxShadow: effects.elevation64,
         pointerEvents: 'auto',
         position: 'absolute',
         display: 'flex',
@@ -251,7 +252,7 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
         overflowX: 'hidden',
         overflowY: 'auto',
         WebkitOverflowScrolling: 'touch',
-        maxHeight: '100%',
+        maxHeight: windowHeight,
         bottom: 0,
         top: 0,
         // (left, right, width) - Properties to be overridden depending on the type of the Panel and the screen breakpoint.
@@ -259,9 +260,6 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
         right: panelMargin.none,
         width: panelWidth.full,
         selectors: {
-          ['@supports (-webkit-overflow-scrolling: touch)']: {
-            maxHeight: windowHeight
-          },
           [HighContrastSelector]: {
             borderLeft: `3px solid ${palette.neutralLight}`,
             borderRight: `3px solid ${palette.neutralLight}`
@@ -285,12 +283,7 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
         maxWidth: '100vw'
       },
       isFooterAtBottom && {
-        height: '100%',
-        selectors: {
-          ['@supports (-webkit-overflow-scrolling: touch)']: {
-            height: windowHeight
-          }
-        }
+        height: windowHeight
       },
       isOpen && isAnimating && !isOnRightSide && AnimationClassNames.slideRightIn40,
       isOpen && isAnimating && isOnRightSide && AnimationClassNames.slideLeftIn40,
@@ -315,21 +308,11 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
         display: 'flex',
         flexDirection: 'column',
         flexGrow: 1,
-        maxHeight: '100%',
-        overflowY: 'hidden',
-        selectors: {
-          ['@supports (-webkit-overflow-scrolling: touch)']: {
-            maxHeight: windowHeight
-          }
-        }
+        maxHeight: windowHeight,
+        overflowY: 'hidden'
       },
       isFooterAtBottom && {
-        height: '100%',
-        selectors: {
-          ['@supports (-webkit-overflow-scrolling: touch)']: {
-            height: windowHeight
-          }
-        }
+        height: windowHeight
       }
     ],
     header: [
@@ -351,7 +334,9 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
       DefaultFontStyles.xLarge,
       {
         color: palette.neutralPrimary,
-        lineHeight: '32px',
+        fontSize: 20, // TODO: after the type ramp gets reevaluated this needs to be changed
+        fontWeight: FontWeights.semibold,
+        lineHeight: '27px',
         margin: 0
       },
       headerClassName
@@ -360,12 +345,7 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
       classNames.scrollableContent,
       {
         overflowY: 'auto',
-        height: '100%',
-        selectors: {
-          ['@supports (-webkit-overflow-scrolling: touch)']: {
-            height: windowHeight
-          }
-        }
+        height: windowHeight
       }
     ],
     content: [
@@ -393,8 +373,8 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
       classNames.footerInner,
       sharedPaddingStyles,
       {
-        paddingBottom: '20px',
-        paddingTop: '20px'
+        paddingBottom: 16,
+        paddingTop: 16
       }
     ]
     // subComponentStyles: {