+++ packages/office-ui-fabric-react/src/components/Stack/examples/Stack.Vertical.Grow.Example.tsx
@@ -1,35 +1,43 @@
 import * as React from 'react';
-import { mergeStyleSets, DefaultPalette, IStackTokens, Stack } from 'office-ui-fabric-react';
+import { Stack } from '../Stack';
+import { IStackStyles, IStackTokens } from '../Stack.types';
+import { IStackItemStyles } from '../StackItem/StackItem.types';
+import { DefaultPalette } from 'office-ui-fabric-react/lib/Styling';
 
 export class VerticalStackGrowExample extends React.Component<{}, {}> {
   public render(): JSX.Element {
-    const styles = mergeStyleSets({
+    const stackStyles: IStackStyles = {
       root: {
         background: DefaultPalette.themeTertiary,
         height: 250
-      },
-
-      item: {
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        background: DefaultPalette.themePrimary,
-        color: DefaultPalette.white
       }
-    });
+    };
+    const stackItemStyles: IStackItemStyles = {
+      root: {
+        alignItems: 'center',
+        background: DefaultPalette.themePrimary,
+        color: DefaultPalette.white,
+        display: 'flex',
+        justifyContent: 'center'
+      }
+    };
 
-    const stackTokens: IStackTokens = { childrenGap: 5 };
+    const outerStackTokens: IStackTokens = { childrenGap: 5 };
+    const innerStackTokens: IStackTokens = {
+      childrenGap: 5,
+      padding: 10
+    };
 
     return (
-      <Stack tokens={stackTokens}>
-        <Stack tokens={stackTokens} padding={10} className={styles.root}>
-          <Stack.Item grow={3} className={styles.item}>
+      <Stack tokens={outerStackTokens}>
+        <Stack styles={stackStyles} tokens={innerStackTokens}>
+          <Stack.Item grow={3} styles={stackItemStyles}>
             Grow is 3
           </Stack.Item>
-          <Stack.Item grow={2} className={styles.item}>
+          <Stack.Item grow={2} styles={stackItemStyles}>
             Grow is 2
           </Stack.Item>
-          <Stack.Item grow className={styles.item}>
+          <Stack.Item grow styles={stackItemStyles}>
             Grow is 1
           </Stack.Item>
         </Stack>