+++ apps/fabric-website/CHANGELOG.md
@@ -1,6 +1,29 @@
 # Change Log - @uifabric/fabric-website
 
-This log was last generated on Fri, 31 May 2019 12:21:27 GMT and should not be manually modified.
+This log was last generated on Wed, 12 Jun 2019 00:42:26 GMT and should not be manually modified.
+
+## 6.19.0
+Wed, 12 Jun 2019 00:42:26 GMT
+
+### Minor changes
+
+- Changes to support slots API refactoring
+
+### Patches
+
+- Update Calendar and DatePicker import paths
+- Update and dedupe React deps.
+- Support slots API changes.
+- Nav, Calendar, and DatePicker move into office-ui-fabric-react
+- Update Router imports
+
+## 6.18.11
+Thu, 06 Jun 2019 12:22:24 GMT
+
+### Patches
+
+- Add redirects to new website
+- Fix uhf style override className.
 
 ## 6.18.10
 Fri, 31 May 2019 12:21:27 GMT